+++ Insight/service/service.py
@@ -142,7 +142,7 @@ class service_module(object):
 
     @classmethod
     def get_version(cls):
-        version_str = 'v1.2.0'
+        version_str = 'v1.3.0'
         return LooseVersion(version_str)
 
     @staticmethod